+++ ErsatzTV.Core/Domain/PlaybackOrder.cs
@@ -6,5 +6,6 @@ public enum PlaybackOrder
     Random = 2,
     Shuffle = 3,
     ShuffleInOrder = 4,
-    MultiEpisodeShuffle = 5
+    MultiEpisodeShuffle = 5,
+    SeasonEpisode = 6
 }